update plugin manager

+++ example/performace/cython/algorithm.pyx
@@ -1,5 +1,5 @@
 def fast_sum(int[:] arr):
-    cdef int i, total = 0
+    cdef long long i, total = 0
     for i in range(len(arr)):
         total += arr[i]
     return total
